Set up change observer _before_ performing change
Prior to this change, the test changed the line ending for a file,
_then_ registered a change observer, and then waited for the change
observer to be called. We're seeing some flakiness with this test
(https://github.com/atom/atom/issues/17326), and I think this order of
operations might be the cause of the flakiness.

To make sure we observer gets called when the line endings change, let's
register the observer _before_ changing the line endings.